% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{ApplyVarianceTransformations}
\alias{ApplyVarianceTransformations}
\title{Perform and visualize different variance stabilization transformations}
\usage{
ApplyVarianceTransformations(dds, outpath)
}
\description{
Perform and visualize different variance stabilization transformations
}
