% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateOutputStructure}
\alias{CreateOutputStructure}
\title{Create output directory structure}
\usage{
CreateOutputStructure(block, level, base, chip = FALSE)
}
\arguments{
\item{block}{String or character vector defining the variable(s) to use to
block for unwanted variance, e.g. batch or technical effects.}

\item{level}{String defining variable of interest.}

\item{base}{String defining the base output path.}

\item{chip}{Boolean indicating whether output is being created for ChIP-seq.}
}
\value{
A named List containing the modified base output path and design
  formula.
}
\description{
\code{CreateRNAOutputStructure} generates the output directories used by
\code{\link{RunDESeq2}}, \code{\link{ProcessDEGs}}, 
\code{\link{RunDiffBind}}, and \code{\link{ProcessDBRs}}.
}
\author{
Jared Andrews
}
