% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateRNAOutputStructure}
\alias{CreateRNAOutputStructure}
\title{Create RNA-seq output directory structure}
\usage{
CreateRNAOutputStructure(block, level, base)
}
\arguments{
\item{block}{String or character vector defining the variable(s) to use to
block for unwanted variance, e.g. batch or technical effects.}

\item{level}{String defining variable of interest.}

\item{base}{String defining the base output path.}
}
\value{
A named List containing the modified base output path and design
}
\description{
\code{CreateRNAOutputStructure} generates the output directories used by
\code{\link{RunDESeq2}} and \code{\link{ProcessDEGs}}.
}
\author{
Jared Andrews
}
