% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEnrichments.R
\name{DoEnrichments}
\alias{DoEnrichments}
\title{Perform GO/Pathway enrichment for DEGs via enrichR}
\usage{
DoEnrichments(samp1, samp2, results, baseout, padj.thresh = 0.01,
  lfc.thresh = 2, top.n = NULL)
}
\description{
Perform GO/Pathway enrichment for DEGs via enrichR
}
\author{
Jared Andrews
}
