% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{GetVarianceTransformations}
\alias{GetVarianceTransformations}
\title{Perform and visualize different variance stabilization transformations}
\usage{
GetVarianceTransformations(dds, outpath)
}
\description{
Perform and visualize different variance stabilization transformations
}
\author{
Jared Andrews
}
