% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotBoxplots}
\alias{PlotBoxplots}
\title{Plot gene counts as box plots}
\usage{
PlotBoxplots(res.list, dds, rld, outpath, padj.thresh, fc.thresh, top.n,
  level)
}
\arguments{
\item{res.list}{Named List containing \linkS4class{DESeqResults} objects for 
all comparisons generated by \code{\link{ProcessDEGs}}.}

\item{dds}{A \linkS4class{DESeqDataSet} object as returned by 
\code{\link[DESeq2]{DESeq}} or \link{RunDESeq2}.}

\item{rld}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{rlog}} transformed counts as returned by
\link{RunDESeq2}.}

\item{outpath}{Path to directory to be used for output.}

\item{padj.thresh}{Number indicating the adjusted p-value 
cutoff to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number indicating the log2 fold-change 
cutoff to be used for determining "significant" differential expression.}

\item{top.n}{Number of differentially expressed genes to create boxplots for, 
ranked by adj. p-value after applying \code{padj.thresh} and 
\code{fc.thresh} thresholds.}

\item{level}{String defining variable of interest.}
}
\description{
Plot gene counts as box plots
}
\author{
Jared Andrews
}
