% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqViz.R
\name{PlotChIPHeatmaps}
\alias{PlotChIPHeatmaps}
\title{Plot ChIP heatmaps}
\usage{
PlotChIPHeatmaps(
  results,
  outpath,
  method,
  breaks,
  colors,
  fdr.thresh = 1,
  fc.thresh = 0,
  consensus = TRUE
)
}
\arguments{
\item{results}{\code{DBA} object as returned by 
\code{\link[DiffBind]{dba.analyze}}.}

\item{outpath}{Path to directory to be used for output.}

\item{method}{Method used for \code{DiffBind} analyses 
(e.g. \code{DBA_DESEQ2}).}

\item{breaks}{Scalar vector of breaks to use for color mapping.}

\item{colors}{Vectors of colors to use for color mapping.}

\item{fdr.thresh}{Number indicating FDR threshold. Peaks greater than
\code{fdr.thresh} will not be included in the heatmap or used for sample
correlations.}

\item{fc.thresh}{Number indicating absolute log fold-change that peaks must 
meet to be included in the heatmaps or used for sample correlations.}

\item{consensus}{Boolean indicating whether creating heatmaps for consensus 
peaks.}
}
\description{
\code{PlotChIPHetmaps} generates signal and correlation heatmaps for a 
\code{DBA} object.
}
\details{
This function will generate heatmaps for each contrast individually as well
as all differentially bound peaks in all samples if \code{consensus = FALSE}.
A table will be saved of all peaks in the combined heatmap, with row order
retained if row clustering is performed.
}
\author{
Jared Andrews
}
