% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqViz.R
\name{PlotChIPPCAs}
\alias{PlotChIPPCAs}
\title{Plot ChIP PCAs}
\usage{
PlotChIPPCAs(
  results,
  outpath,
  method,
  fdr.thresh = 1,
  fc.thresh = 0,
  consensus = TRUE
)
}
\arguments{
\item{results}{\code{DBA} object as returned by 
\code{\link[DiffBind]{dba.analyze}}.}

\item{outpath}{Path to directory to be used for output.}

\item{method}{Method used for \code{DiffBind} analyses 
(e.g. \code{DBA_DESEQ2}).}

\item{fdr.thresh}{Number indicating FDR threshold. Peaks greater than
\code{fdr.thresh} will not be used for PCAs.}

\item{fc.thresh}{Number indicating absolute log fold-change that peaks must 
meet to be used for the PCA.}

\item{consensus}{Boolean indicating whether PCAs are for consensus peaks.}
}
\description{
\code{PlotChIPPCAs} generates PCA plots for all contrasts found in 
\code{results}. If \code{consensus = TRUE}, it will also generate one with
all samples and all peaks.
}
\author{
Jared Andrews
}
