% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotDEGPCAs}
\alias{PlotDEGPCAs}
\title{Plot PCAs from variance stabilized counts for differentially expressed genes}
\usage{
PlotDEGPCAs(res.list, rld, vsd, outpath, level, plot.annos, padj.thresh,
  fc.thresh)
}
\description{
Plot PCAs from variance stabilized counts for differentially expressed genes
}
\author{
Jared Andrews
}
