% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{PlotEDAPCAs}
\alias{PlotEDAPCAs}
\title{Plot PCAs from variance stabilized counts for EDA}
\usage{
PlotEDAPCAs(rld, vsd, outpath, level, plot.annos)
}
\description{
Plot PCAs from variance stabilized counts for EDA
}
\author{
Jared Andrews
}
