% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEnrichments.R
\name{PlotEnrichments}
\alias{PlotEnrichments}
\title{Perform GO/Pathway enrichment for DEGs via enrichR}
\usage{
PlotEnrichments(res.list, outpath, padj.thresh, fc.thresh)
}
\arguments{
\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential expression.}
}
\description{
Perform GO/Pathway enrichment for DEGs via enrichR
}
\author{
Jared Andrews
}
