% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotRNACombinedHeatmaps}
\alias{PlotRNACombinedHeatmaps}
\title{Generate gene heatmaps for all DEGs from a list of DESeq2 results}
\usage{
PlotRNACombinedHeatmaps(res.list, rld, vsd, outpath, padj.thresh,
  fc.thresh, plot.annos)
}
\arguments{
\item{res.list}{Named List containing \linkS4class{DESeqResults} objects for 
all comparisons generated by \code{\link{ProcessDEGs}}.}

\item{rld}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{rlog}} transformed counts as returned by
\link{RunDESeq2}.}

\item{vsd}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{vst}} transformed counts as returned by
\link{RunDESeq2}.}

\item{outpath}{Path to directory to be used for output.}

\item{padj.thresh}{Number indicating the adjusted p-value 
cutoff to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number indicating the log2 fold-change 
cutoff to be used for determining "significant" differential expression.}

\item{plot.annos}{String or character vector defining the column(s) to use to 
annotate figures.}
}
\description{
Generate gene heatmaps for all DEGs from a list of DESeq2 results
}
\author{
Jared Andrews
}
