% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{PlotRNASampleDistances}
\alias{PlotRNASampleDistances}
\title{Visualize sample distances from variance stabilized counts}
\usage{
PlotRNASampleDistances(rld, vsd, outpath, level, plot.annos)
}
\arguments{
\item{rld}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{rlog}} transformed counts as returned by
\link{PlotRNAVarianceTransformations}.}

\item{vsd}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{vst}} transformed counts as returned by
\link{PlotRNAVarianceTransformations}.}

\item{outpath}{Path to output file.}

\item{level}{String defining variable of interest.}

\item{plot.annos}{String or character vector defining the column(s) in 
\code{samplesheet} to use to annotate figures.}
}
\description{
Visualize sample distances from variance stabilized counts
}
\author{
Jared Andrews
}
