% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{PlotRNAVarianceTransformations}
\alias{PlotRNAVarianceTransformations}
\title{Perform and visualize different variance stabilization transformations}
\usage{
PlotRNAVarianceTransformations(dds, outpath)
}
\arguments{
\item{dds}{A \linkS4class{DESeqDataSet} object.}

\item{outpath}{Path to output file.}
}
\description{
Perform and visualize different variance stabilization transformations
}
\author{
Jared Andrews
}
