% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotRNAVolcanoes}
\alias{PlotRNAVolcanoes}
\title{Create volcano plots from a list of DESeq2 Results objects}
\usage{
PlotRNAVolcanoes(res.list, dds, outpath, padj.thresh, fc.thresh,
  n.labels = 10, use.labels = TRUE)
}
\arguments{
\item{res.list}{Named List containing \linkS4class{DESeqResults} objects for 
all comparisons generated by \code{\link{ProcessDEGs}}.}

\item{dds}{A \linkS4class{DESeqDataSet} object as returned by 
\code{\link[DESeq2]{DESeq}}, \link{RunDESeq2}, or \link{ProcessDEGs}.}

\item{outpath}{Path to directory to be used for output.}

\item{padj.thresh}{Number indicating the adjusted p-value 
cutoff to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number indicating the log2 fold-change 
cutoff to be used for determining "significant" differential expression.}

\item{n.labels}{Number of genes to label on the volcano plot, ranked by 
adjusted p-value.}

\item{use.labels}{Boolean indicating whether labels should be added to plot.}
}
\description{
Create volcano plots from a list of DESeq2 Results objects
}
\author{
Jared Andrews
}
