% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{PlotVarianceTransformations}
\alias{PlotVarianceTransformations}
\title{Perform and visualize different variance stabilization transformations}
\usage{
PlotVarianceTransformations(dds, outpath)
}
\description{
Perform and visualize different variance stabilization transformations
}
\author{
Jared Andrews
}
