% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotVolcanoes}
\alias{PlotVolcanoes}
\title{Create volcano plots from results lists for all genes}
\usage{
PlotVolcanoes(res.list, dds, outpath, padj.thresh, fc.thresh)
}
\description{
Create volcano plots from results lists for all genes
}
\author{
Jared Andrews
}
