% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{RunDESeq2}
\alias{RunDESeq2}
\title{Run DESeq2 analyses}
\usage{
RunDESeq2(outpath, quants.path, samplesheet, tx2gene, level, g1, g2,
  padj.thresh = 0.05, fc.thresh = 2, plot.annos = NULL,
  block = NULL, read.filt = 10)
}
\arguments{
\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{quants.path}{Path to directory containing a directory for each sample
with a salmon-generated \code{quant.sf} file inside.}

\item{samplesheet}{Path to samplesheet containing sample metadata.}

\item{tx2gene}{Path to file with transcript IDs in first column and gene 
identifiers in second. Used for gene-level summarization.}

\item{level}{String defining variable of interest.}

\item{g1}{String indicating group in \code{level} to compare to \code{g2}.}

\item{g2}{String indicating group in \code{level} to compare to \code{g1}.}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential expression.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential expression.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{plot.annos}{String or character vector defining the column(s) in 
\code{samplesheet} to use to annotate figures.}

\item{block}{String or character vector defining the column(s) in 
\code{samplesheet} to use to block for unwanted variance, e.g. batch or 
technical effects.}

\item{read.filt}{Number indicating read threshold. Genes with fewer counts 
than this number summed across all samples will be removed from the
analysis.}
}
\value{
Named List containing a \linkS4class{DESeqDataSet} object from 
  running \code{\link[DESeq2]{DESeq}}, a 
  \linkS4class{RangedSummarizedExperiment} object from running 
  \code{\link[DESeq2]{rlog}}, and a \linkS4class{RangedSummarizedExperiment} 
  object from running \code{\link[DESeq2]{vst}}.
}
\description{
This function performs a high-level differential gene expression analysis
}
\seealso{
\code{\link[DESeq2]{DESeq}}, for more about differential expression analysis.
}
