% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEDA.R
\name{RunVarianceTransformations}
\alias{RunVarianceTransformations}
\title{Perform and visualize different variance stabilization transformations}
\usage{
RunVarianceTransformations(dds, outpath)
}
\description{
Perform and visualize different variance stabilization transformations
}
