% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SaveRNAResults}
\alias{SaveRNAResults}
\title{Save comparison results}
\usage{
SaveRNAResults(res.list, dds, outpath)
}
\arguments{
\item{res.list}{Named List containing \linkS4class{DESeqResults} objects for 
all comparisons generated by \code{\link{ProcessDEGs}}.}

\item{dds}{A \linkS4class{DESeqDataSet} object as returned by 
\code{\link[DESeq2]{DESeq}}.}

\item{outpath}{Path to directory to be used for output.}
}
\description{
\code{SaveRNAResults} saves the results for each comparison in \code{res.list} 
along with the gene counts.
}
\author{
Jared Andrews
}
