% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateOutputStructure}
\alias{CreateOutputStructure}
\title{Create output directory structure}
\usage{
CreateOutputStructure(block, level, base)
}
\arguments{
\item{block}{String or character vector defining the variable(s) to use to 
block for unwanted variance, e.g. batch or technical effects.}

\item{level}{String defining variable of interest.}

\item{base}{String defining the base output path.}
}
\value{
A named List containing the modified base output path and design
}
\description{
\code{CreateOutputStructure} generates the output directories used by
\code{\link{RunDESeq2}}.
}
