% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotBoxplots}
\alias{PlotBoxplots}
\title{Plot gene counts as box plots}
\usage{
PlotBoxplots(res.list, dds, rld, outpath, padj.thresh, fc.thresh, top.n,
  level)
}
\description{
Plot gene counts as box plots
}
\author{
Jared Andrews
}
