% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqViz.R
\name{PlotChIPAnnos}
\alias{PlotChIPAnnos}
\title{Plot ChIP annotation breakdowns}
\usage{
PlotChIPAnnos(peak.anno, outpath, consensus = TRUE, comp = NULL,
  fc = NULL, fdr = NULL)
}
\arguments{
\item{peak.anno}{A named list of \linkS4class{csAnno} objects with annotated
peaks as returned by \code{\link[ChIPseeker]{annotatePeak}}. If 
\code{consensus = TRUE}, only one is expected, otherwise there should be
three (All DBRs, group1 up, group2 up).}

\item{outpath}{Path to directory to be used for output.}

\item{consensus}{Boolean indicating if consensus peaks are provided.}

\item{comp}{Name of comparison, for file naming.}

\item{fc}{Fold change threshold, for file naming.}

\item{fdr}{FDR threshold, for file naming.}
}
\description{
\code{PlotChIPAnnos} generates various plots for all annotated peaksets 
(as generated by \code{\link[ChIPseeker]{annotatePeak}}) found in 
\code{peak.anno}.
}
\author{
Jared Andrews
}
