% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEnrichments.R
\name{PlotEnrichments}
\alias{PlotEnrichments}
\title{Perform GO/Pathway enrichment for DEGs via enrichR}
\usage{
PlotEnrichments(res.list, outpath, padj.thresh, fc.thresh,
  libraries = c("GO_Molecular_Function_2018",
  "GO_Cellular_Component_2018", "GO_Biological_Process_2018",
  "KEGG_2019_Human", "Reactome_2016", "BioCarta_2016", "Panther_2016"))
}
\arguments{
\item{res.list}{Named List containing \linkS4class{DESeqResults} objects for 
all comparisons generated by \code{\link{ProcessDEGs}}.}

\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{padj.thresh}{Number indicating the adjusted p-value 
cutoff to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number indicating the log2 fold-change 
cutoff to be used for determining "significant" differential expression.}

\item{libraries}{A vector of valid \code{enrichR} libraries to test the genes 
against.}
}
\description{
\code{PlotEnrichments} performs and plots enrichment analyses for
}
\author{
Jared Andrews
}
