% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqEnrichments.R
\name{PlotEnrichments}
\alias{PlotEnrichments}
\title{Perform GO/Pathway enrichment for DEGs via enrichR}
\usage{
PlotEnrichments(res.list, outpath, padj.thresh, fc.thresh,
  libraries = c("GO_Molecular_Function_2018",
  "GO_Cellular_Component_2018", "GO_Biological_Process_2018",
  "KEGG_2019_Human", "Reactome_2016", "BioCarta_2016", "Panther_2016"))
}
\arguments{
\item{res.list}{Named List containing \linkS4class{DESeqResults} objects for 
all comparisons generated by \code{\link{ProcessDEGs}}.}

\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{padj.thresh}{Number indicating the adjusted p-value 
cutoff to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number indicating the log2 fold-change 
cutoff to be used for determining "significant" differential expression.}

\item{libraries}{A vector of valid \code{enrichR} libraries to test the genes 
against.}
}
\description{
\code{PlotEnrichments} performs and plots enrichment analyses for a list of 
\linkS4class{DESeqResults} objects as generated by \link{RunDESeq2} or
\link{ProcessDEGs}. It queries the enrichr web server to do so. Results are
saved both as tables and plots.
}
\details{
Available libraries can be viewed with \code{listEnrichrDbs} from the
\code{enrichR} package.

The enrichr web server can be found at 
\url{https://amp.pharm.mssm.edu/Enrichr/}. If you use this function, you 
should be sure to cite the original authors.
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/27141961}{Kuleshov MV, 
  Jones MR, Rouillard AD, et al. Enrichr: a comprehensive gene set enrichment 
  analysis web server 2016 update. Nucleic Acids Research. 2016; gkw377}
}
\author{
Jared Andrews
}
