% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{PlotEnrichments}
\alias{PlotEnrichments}
\title{Perform GO/Pathway enrichment for DEGs or DBRs via enrichR}
\usage{
PlotEnrichments(
  res.list,
  outpath,
  padj.thresh,
  fc.thresh,
  libraries,
  chip = FALSE
)
}
\arguments{
\item{res.list}{Either:
\itemize{
  \item A named list containing \linkS4class{DESeqResults} objects for 
    all comparisons generated by \code{\link{ProcessDEGs}}, or
  \item A named list containing two or three \linkS4class{csAnno} objects 
    for a given comparison generated by \code{\link{ProcessDBRs}} if 
    \code{chip = TRUE}. First must contain all DBRs, second/third contain
    DBRs up in one group versus the other.
}}

\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{padj.thresh}{Number indicating the adjusted p-value 
cutoff to be used for determining "significant" differential expression
or differential binding (if \code{chip = TRUE}).}

\item{fc.thresh}{Number indicating the log2 fold-change 
cutoff to be used for determining "significant" differential expression.
or differential binding (if \code{chip = TRUE}).}

\item{libraries}{A vector of valid \code{enrichR} libraries to test the genes 
against.}

\item{chip}{Boolean indicating whether \code{res.list} is the results from
\code{\link[DiffBind]{dba.analyze}}.}
}
\description{
\code{PlotEnrichments} performs and plots enrichment analyses for a list of 
\linkS4class{DESeqResults} objects as generated by \code{\link{RunDESeq2}} or
\code{\link{ProcessDEGs}} or a list of \code{data.frames} as generated by
\code{\link{RunDiffBind}} or \code{\link{ProcessDBRs}}. It queries the 
enrichr web server to do so. Results are saved both as tables and plots.
}
\details{
Available libraries can be viewed with \code{\link[enrichR]{listEnrichrDbs}} 
from the \code{enrichR} package.

The enrichr web server can be found at 
\url{https://amp.pharm.mssm.edu/Enrichr/}. If you use this function, you 
should be sure to cite the original authors.
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/27141961}{Kuleshov MV, 
  Jones MR, Rouillard AD, et al. Enrichr: a comprehensive gene set enrichment 
  analysis web server 2016 update. Nucleic Acids Research. 2016; gkw377}
}
\seealso{
\code{\link[EZscRNA]{RunEnrichr}}, \code{\link[EZscRNA]{VizEnrichments}}, and
\code{\link[enrichR]{enrichr}} for additional enrichment options. 

\code{\link{ProcessDEGs}} for generating \linkS4class{DESeqResults} objects.
\code{\link{ProcessDBRs}}, for generating \linkS4class{csAnno} objects.
}
\author{
Jared Andrews
}
