% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqViz.R
\name{PlotHeatmaps}
\alias{PlotHeatmaps}
\title{Generate gene heatmaps for a list of DESeq2 results}
\usage{
PlotHeatmaps(res.list, dds, rld, vsd, level, outpath, padj.thresh,
  fc.thresh, plot.annos = NULL)
}
\arguments{
\item{level}{String defining variable of interest.}

\item{outpath}{Path to directory to be used for output.}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential expression.}

\item{plot.annos}{String or character vector defining the column(s) to use to 
annotate figures.}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential expression.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential expression.}
}
\description{
Generate gene heatmaps for a list of DESeq2 results
}
\author{
Jared Andrews
}
