% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseq.R
\name{ProcessDBRs}
\alias{ProcessDBRs}
\title{Extract, visualize, and save DBRs}
\usage{
ProcessDBRs(results, rblock, outpath, level, padj.thresh = 0.05,
  fc.thresh = 2, breaks = c(seq(-3, -1, length = 250), seq(-1, -0.1,
  length = 250), seq(-0.1, 0.1, length = 1), seq(0.1, 1, length = 250),
  seq(1, 3, length = 250)), heatmap.colors = NULL,
  heatmap.preset = NULL, plot.enrich = TRUE,
  enrich.libs = c("GO_Molecular_Function_2018",
  "GO_Cellular_Component_2018", "GO_Biological_Process_2018",
  "KEGG_2019_Human", "Reactome_2016", "BioCarta_2016", "Panther_2016"))
}
\arguments{
\item{results}{A \linkS4class{DESeqDataSet} object as returned by 
\code{\link[DESeq2]{DESeq}}.}

\item{rblock}{Boolean indicating whether a blocking variable was used or not.}

\item{outpath}{Path to directory to be used for output.}

\item{level}{String defining variable of interest in samplesheet.
Must be one of: "Treatment", "Condition", "Tissue", or "Factor".}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential binding.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential binding.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{breaks}{Vector of sequences to be used as breaks for signal heatmaps.}

\item{heatmap.colors}{Character vector containing custom colors to use for 
heatmaps in hex (e.g. \code{c("#053061", "#f5f5f5", "#67001f")}).}

\item{heatmap.preset}{String indicating which of the color presets to use in
  heatmaps.

  Available presets (low to high) are:
  \itemize{
     \item "BuRd" Blue to red.
     \item "OrPu" Orange to purple. 
     \item "BrTe" Brown to teal.
     \item "PuGr" Purple to green.
  }}

\item{plot.enrich}{Boolean indicating whether enrichment analyses for DBRs 
should be run and plotted for each comparison.}

\item{enrich.libs}{A vector of valid \code{enrichR} libraries to test the 
  genes against.

  Available libraries can be viewed with 
  \code{\link[enrichR]{listEnrichrDbs}} from the \code{enrichR} package.}
}
\value{
Named List containing a list named 'res.list' containing 
  \linkS4class{DESeqResults} objects for all comparisons generated by 
  \code{\link{ProcessDEGs}}, a \linkS4class{DESeqDataSet} object named 'dds'
  from running \code{\link[DESeq2]{DESeq}}, a 
  \linkS4class{RangedSummarizedExperiment} object named 'rld' from running 
  \code{\link[DESeq2]{rlog}}, and a \linkS4class{RangedSummarizedExperiment} 
  object from running \code{\link[DESeq2]{vst}} named 'vsd'.
}
\description{
\code{ProcessDBRs} wraps several plotting functions to generate figures 
specifically for the differentially bound regions between each possible
comparison. It is called by \link{RunDiffBind} but can also be re-run with 
the \link{RunDiffBind} output or its own output if you want to save time and 
don't need to generate all of the exploratory data analysis figures again.
}
\details{
This function will generate many figures in addition to saving the results.
}
\seealso{
\code{\link{RunDiffBind}}, for generating input for this function.
}
\author{
Jared Andrews
}
