% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{ProcessDEGs}
\alias{ProcessDEGs}
\title{Extract, visualize, and save DEG lists}
\usage{
ProcessDEGs(dds, rld, vsd, outpath, level, plot.annos,
  padj.thresh = 0.05, fc.thresh = 2, plot.box = TRUE,
  plot.enrich = TRUE, enrich.libs = c("GO_Molecular_Function_2018",
  "GO_Cellular_Component_2018", "GO_Biological_Process_2018",
  "KEGG_2019_Human", "Reactome_2016", "BioCarta_2016", "Panther_2016"),
  top.n = 100)
}
\arguments{
\item{dds}{A \linkS4class{DESeqDataSet} object as returned by 
\code{\link[DESeq2]{DESeq}}.}

\item{rld}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{rlog}} transformed counts as returned by
\link{RunDESeq2}.}

\item{vsd}{A \linkS4class{RangedSummarizedExperiment} object of 
\code{\link[DESeq2]{vst}} transformed counts as returned by
\link{RunDESeq2}.}

\item{outpath}{Path to directory to be used for output.}

\item{level}{String defining variable of interest.}

\item{plot.annos}{String or character vector defining the column(s) in 
\code{samplesheet} to use to annotate figures.}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential expression.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential expression.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{plot.box}{Boolean indicating whether box plots for DEGs should be 
created for each comparison. If so, the \code{top.n} genes will be plotted.}

\item{plot.enrich}{Boolean indicating whether enrichment analyses for DEGs 
should be run and plotted for each comparison.}

\item{enrich.libs}{A vector of valid \code{enrichR} libraries to test the 
  genes against.

  Available libraries can be viewed with 
  \code{\link[enrichR]{listEnrichrDbs}} from the \code{enrichR} package.}

\item{top.n}{Number of differentially expressed genes to create boxplots for, 
ranked by adj. p-value after applying \code{padj.thresh} and 
\code{fc.thresh} thresholds. If multiple thresholds are provided, the 
lowest fold-change and highest adj. p-value thresholds will be used.}
}
\value{
Named List containing a list named 'res.list' containing named lists
  for each p-value threshold containing
  \linkS4class{DESeqResults} objects for all comparisons generated by 
  \code{\link{ProcessDEGs}}, a \linkS4class{DESeqDataSet} object named 'dds'
  from running \code{\link[DESeq2]{DESeq}}, a 
  \linkS4class{RangedSummarizedExperiment} object named 'rld' from running 
  \code{\link[DESeq2]{rlog}}, and a \linkS4class{RangedSummarizedExperiment} 
  object from running \code{\link[DESeq2]{vst}} named 'vsd'.
}
\description{
\code{ProcessDEGs} wraps several plotting functions to generate figures 
specifically for the differentially expressed genes between each possible
comparison.
}
\details{
\code{ProcessDEGs} is called by \link{RunDESeq2} but can also be re-run with 
the \link{RunDESeq2} output or its own output if you want to save time and 
don't need to generate all of the exploratory data analysis figures again.

This function will generate many figures in addition to saving the counts
and results tables.
}
\seealso{
\code{\link{RunDESeq2}}, for generating input for this function.
}
\author{
Jared Andrews
}
