% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqEDA.R
\name{RunChIPQC}
\alias{RunChIPQC}
\title{Run ChIPQC on a sample sheet}
\usage{
RunChIPQC(outpath, samplesheet, chromosomes = "chr18")
}
\arguments{
\item{outpath}{Path to directory to be used for output.}

\item{samplesheet}{Path to samplesheet containing sample metadata.}

\item{chromosomes}{String or character vector indicating chromosomes to be 
used for QC.}
}
\value{
Dataframe containing QC metrics for each sample, including 
  cross-correlation scores, fraction of reads in peaks, fragment length, etc.
}
\description{
This function simply runs \code{\link[ChIPQC]{ChIPQC}} on a sample sheet and
returns QC metrics. It generates a consensus peakset between all sets by 
retaining and merging all peaks that overlap in at least two samples before
defining the fraction of reads in peaks and other metrics.
}
\details{
A basic plot showing overlap between peaksets will also be generated, which 
can be useful for defining how many samples a peak should be found in to be
included in the consensus peakset as defined in \code{\link{RunDiffBind}}.
}
\seealso{
\code{\link[ChIPQC]{ChIPQC}}
}
\author{
Jared Andrews
}
