% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{RunDESeq2}
\alias{RunDESeq2}
\title{Run DESeq2 analyses}
\usage{
RunDESeq2(outpath, quants.path, samplesheet, tx2gene, level,
  padj.thresh = 0.05, fc.thresh = 2, plot.annos = NULL,
  plot.box = TRUE, plot.enrich = TRUE, top.n = 100, block = NULL,
  count.filt = 10)
}
\arguments{
\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{quants.path}{Path to directory containing a directory for each sample
with a salmon-generated \code{quant.sf} file inside.}

\item{samplesheet}{Path to samplesheet containing sample metadata.}

\item{tx2gene}{Path to file with transcript IDs in first column and gene 
identifiers in second. Used for gene-level summarization.}

\item{level}{String defining variable of interest.}

\item{padj.thresh}{Number or numeric scalar indicating the adjusted p-value 
cutoff(s) to be used for determining "significant" differential expression.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential expression.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{plot.annos}{String or character vector defining the column(s) in 
\code{samplesheet} to use to annotate figures.}

\item{plot.box}{Boolean indicating whether box plots for DEGs should be 
created for each comparison. If so, the \code{top.n} genes will be plotted.
This step is quite time-consuming with many genes.}

\item{plot.enrich}{Boolean indicating whether enrichment analyses for DEGs 
should be run and plotted for each comparison.}

\item{top.n}{Number of differentially expressed genes to create boxplots for, 
ranked by adj. p-value after applying \code{padj.thresh} and 
\code{fc.thresh} thresholds. If multiple thresholds are provided, the 
lowest fold-change and highest adj. p-value thresholds will be used.}

\item{block}{String or character vector defining the column(s) in 
\code{samplesheet} to use to block for unwanted variance, e.g. batch or 
technical effects.}

\item{count.filt}{Number indicating read threshold. Genes with fewer counts 
than this number summed across all samples will be removed from the
analysis.}
}
\value{
Named List containing a list named 'res.list' containing 
  \linkS4class{DESeqResults} objects for all comparisons generated by 
  \code{\link{ProcessDEGs}}, a \linkS4class{DESeqDataSet} object named 'dds'
  from running \code{\link[DESeq2]{DESeq}}, a 
  \linkS4class{RangedSummarizedExperiment} object named 'rld' from running 
  \code{\link[DESeq2]{rlog}}, and a \linkS4class{RangedSummarizedExperiment} 
  object from running \code{\link[DESeq2]{vst}} named 'vsd'.
}
\description{
\code{RunDESeq2} performs a high-level differential gene expression analysis 
with \code{DESeq2}. It, along with \link{ProcessDEGs}, form the crux of the
RNA-seq portion of this package.
}
\details{
The default parameters should be an adequate starting place for most users,
but lazy folks can provide multiple thresholds to \code{padj.thresh}
and/or \code{fc.thresh} if they aren't sure how stringent or lenient they 
need to be with their data.

It's generally best to provide an empty directory as the output path, as
several directories will be generated.
}
\seealso{
\code{\link[DESeq2]{DESeq}}, for more about differential expression analysis.
\code{\link{ProcessDEGs}}, for analyzing and visualizing the results.
}
\author{
Jared Andrews
}
