% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseq.R
\name{RunDiffBind}
\alias{RunDiffBind}
\title{Run DiffBind analysis on a samplesheet}
\usage{
RunDiffBind(outpath, samplesheet, txdb, dba = NULL,
  level = c("Treatment", "Condition", "Tissue", "Factor"), se = NULL,
  fdr.thresh = 0.05, fc.thresh = 2, block = NULL,
  heatmap.colors = NULL, heatmap.preset = NULL, reverse = FALSE,
  n.consensus = 2, breaks = c(seq(-3, -1.0001, length = 250), seq(-1,
  -0.1, length = 250), seq(-0.0999, 0.0999, length = 1), seq(0.1, 1, length
  = 250), seq(1.0001, 3, length = 250)), plot.enrich = TRUE,
  enrich.libs = c("GO_Molecular_Function_2018",
  "GO_Cellular_Component_2018", "GO_Biological_Process_2018",
  "KEGG_2019_Human", "Reactome_2016", "BioCarta_2016", "Panther_2016"),
  promoters = c(-2000, 2000), method = c("DESeq2", "edgeR"),
  scale.full = TRUE)
}
\arguments{
\item{outpath}{Path to directory to be used for output. Additional 
directories will be generated within this folder.}

\item{samplesheet}{Path to samplesheet containing sample metadata.}

\item{txdb}{\code{TxDb} object to use for annotation.}

\item{dba}{DBA object as returned by \code{\link[DiffBind]{dba.count}},
\code{\link[DiffBind]{dba.analyze}} or from this function itself. If 
provided, \code{samplesheet} and \code{n.consensus} are ignored.}

\item{level}{String defining variable of interest from \code{samplesheet}. 
Must be one of: "Treatment", "Condition", "Tissue", or "Factor".}

\item{se}{Path to file containing consensus SEs, which will be used to
to annotate whether individual peaks fall within an SE or not.}

\item{fdr.thresh}{Number or numeric scalar indicating the false discovery 
rate (FDR) cutoff(s) to be used for determining "significant" differential 
binding. If multiple are given, multiple tables/plots will be generated 
using all combinations of \code{fdr.thresh} and \code{fc.thresh}.}

\item{fc.thresh}{Number or numeric scalar indicating the log2 fold-change 
cutoff(s) to be used for determining "significant" differential binding.
If multiple are given, multiple tables/plots will be generated using all 
combinations of \code{padj.thresh} and \code{fc.thresh}.}

\item{block}{String or character vector defining the column(s) in 
\code{samplesheet} to use to block for unwanted variance, e.g. batch or 
technical effects. Must be one of: "Treatment", "Condition", "Tissue", 
or "Factor".}

\item{heatmap.colors}{Character vector containing custom colors to use for 
heatmaps in hex (e.g. \code{c("#053061", "#f5f5f5", "#67001f")}).}

\item{heatmap.preset}{String indicating which of the color presets to use in
  heatmaps.

  Available presets (low to high) are:
  \itemize{
     \item "BuRd" Blue to red.
     \item "OrPu" Orange to purple. 
     \item "BrTe" Brown to teal.
     \item "PuGr" Purple to green.
     \item "BuOr" Sea blue to orange.
  }}

\item{reverse}{Boolean indicating whether to flip heatmap color scheme 
(high color will become low, etc).}

\item{n.consensus}{Number of samples in which peaks must overlap for the 
peaks to be merged and included in the consensus peak set.}

\item{breaks}{Vector of sequences to be used as breaks for signal heatmaps.}

\item{plot.enrich}{Boolean indicating whether enrichment analyses for DBRs 
should be run and plotted for each comparison.}

\item{enrich.libs}{Vector of valid \code{enrichR} libraries to test the 
  genes against.
  

  Available libraries can be viewed with 
  \code{\link[enrichR]{listEnrichrDbs}} from the \code{enrichR} package.}

\item{promoters}{Scalar vector containing how many basepairs up and 
downstream of the TSS should be used to define gene promoters.}

\item{method}{String indicating method to be used for differential expression 
analysis. Can be "DESeq2" or "edgeR".}

\item{scale.full}{Boolean indicating whether the full library size (total 
number of reads) for each sample is used for scaling normalization. If
\code{FALSE}, the total number of reads present in the peaks for each 
sample is used (preferable if overall binding levels are expected to be 
similar between samples).}
}
\value{
A \code{DBA} object from \code{\link[DiffBind]{dba.analyze}}.
}
\description{
\code{RunDiffBind} performs a high-level differential binding analysis 
with \code{DiffBind}. It, along with \link{ProcessDBRs}, form the crux of the
ChIP-seq portion of this package.
}
\details{
The default parameters should be an adequate starting place for most users,
but lazy folks can provide multiple thresholds to \code{fdr.thresh}
and/or \code{fc.thresh} if they aren't sure how stringent or lenient they 
need to be with their data.

Providing the resulting DBA object as input to this function can be useful 
when running multiple times with different \code{level}s and \code{block}s 
or thresholds, as it skips bam loading, which is by far the most 
time-intensive part.

It's generally best to provide an empty directory as the output path, as
several directories will be generated.
}
\seealso{
\code{\link[DiffBind]{dba}}, \code{\link[DiffBind]{dba.count}}, 
\code{\link[DiffBind]{dba.contrast}}, \code{\link[DiffBind]{dba.analyze}},
\code{\link[DiffBind]{dba.report}} for more about ChIP-seq differential
binding analysis.

\code{\link{ProcessDBRs}}, for analyzing and visualizing the results.
}
\author{
Jared Andrews
}
