% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SaveResults}
\alias{SaveResults}
\title{Save results}
\usage{
SaveResults(results, outpath, dds = NULL, chip = FALSE,
  method = NULL, promoters = NULL, se = NULL, txdb = NULL)
}
\arguments{
\item{results}{Either:
\itemize{
  \item a named list containing \linkS4class{DESeqResults} objects for 
    all comparisons generated by \code{\link{ProcessDEGs}}, or
  \item a \code{DBA} object from \code{\link[DiffBind]{dba.analyze}}.
}}

\item{outpath}{Path to directory to be used for output.}

\item{dds}{A \linkS4class{DESeqDataSet} object as returned by 
\code{\link[DESeq2]{DESeq}}.}

\item{chip}{Boolean indicating whether \code{results} are from ChIP-seq 
analysis.}

\item{method}{Method used for differential binding e.g. \code{DBA_DESEQ2}.

  Do not quote this parameter, it is a global variable from \code{DiffBind}.
  If a block was used, be sure to include that 
  (e.g. \code{DBA_DESEQ2_BLOCK}).}

\item{promoters}{Scalar vector containing how many basepairs up and 
downstream of the TSS should be used to define gene promoters.}

\item{se}{Dataframe containing location informatin for super enhancers.}

\item{txdb}{\code{TxDb} object to use for annotation.}
}
\description{
\code{SaveResults} saves the results for each comparison in \code{results}. 
If \code{chip = FALSE}, it will also save normalized gene counts.
}
\author{
Jared Andrews
}
